Disabled opModes && renamed autonomous
