--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/classes/enums/TSEPos.java
@@ -1,7 +1,0 @@
-package org.firstinspires.ftc.teamcode.classes.enums;
-
-public enum TSEPos{
-    TOP,
-    MID,
-    BOT
-}
