--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/classes/enums/TurnDirection.java
@@ -1,6 +1,0 @@
-package org.firstinspires.ftc.teamcode.classes.enums;
-
-public enum TurnDirection {
-    LEFT,
-    RIGHT
-}
