+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/enums/TSEPos.java
@@ -1,0 +1,7 @@
+package org.firstinspires.ftc.teamcode.enums;
+
+public enum TSEPos{
+    TOP,
+    MID,
+    BOT
+}
