+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/enums/TurnDirection.java
@@ -1,0 +1,6 @@
+package org.firstinspires.ftc.teamcode.enums;
+
+public enum TurnDirection {
+    LEFT,
+    RIGHT
+}
