+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/test/MotorEncoderTest.java
@@ -1,5 +1,6 @@
 package org.firstinspires.ftc.teamcode.test;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.CRServo;
@@ -7,16 +8,10 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorEx;
 import com.qualcomm.robotcore.hardware.Servo;
 
+
 @TeleOp
 public class MotorEncoderTest extends OpMode {
     DcMotor motorFrontLeft, motorBackLeft, motorFrontRight, motorBackRight;
-    DcMotorEx arm;
-    Servo servoArm, outtake, grabber;
-    DcMotor intake;
-    CRServo spinner;
-    double armPower = 0.3;
-    int armPos = 0;
-    double spinPower = 1.0;
 
     @Override
     public void init() {
