+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/test/dispenserTest.java
@@ -1,9 +1,11 @@
 package org.firstinspires.ftc.teamcode.test;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.Servo;
 
+@Disabled
 @TeleOp
 public class dispenserTest extends OpMode {
     Servo dispenser;
